+++ src/dsp/sony_s_dsp_gaussian.cpp
@@ -19,65 +19,6 @@
 //
 
 #include "sony_s_dsp_gaussian.hpp"
-#include <algorithm>
-#include <cstddef>
-#include <limits>
-
-/// Clamp an integer to a 16-bit value.
-///
-/// @param n a 32-bit integer value to clip
-/// @returns n clipped to a 16-bit value [-32768, 32767]
-///
-inline int clamp_16(int n) {
-    const int lower = std::numeric_limits<int16_t>::min();
-    const int upper = std::numeric_limits<int16_t>::max();
-    return std::max(lower, std::min(n, upper));
-}
-
-int16_t Sony_S_DSP_Gaussian::run(int16_t input) {
-    // VoiceState& voice = voice_state;
-    // cast the input to 32-bit to do maths
-    int delta = input;
-    // One, two and three point IIR filters
-    int smp1 = voice_state.samples[0];
-    int smp2 = voice_state.samples[1];
-    if (voice_state.filter1) {
-        delta += smp1;
-        delta -= smp2 >> 1;
-        if (!voice_state.filter2) {
-            delta += (-smp1 - (smp1 >> 1)) >> 5;
-            delta += smp2 >> 5;
-        } else {
-            delta += (-smp1 * 13) >> 7;
-            delta += (smp2 + (smp2 >> 1)) >> 4;
-        }
-    } else if (voice_state.filter2) {
-        delta += smp1 >> 1;
-        delta += (-smp1) >> 5;
-    }
-    // update sample history
-    voice_state.samples[3] = voice_state.samples[2];
-    voice_state.samples[2] = smp2;
-    voice_state.samples[1] = smp1;
-    voice_state.samples[0] = 2 * clamp_16(delta);
-
-    // get the 14-bit frequency value
-    // TODO:
-    // int rate = 0x3FFF & ((raw_voice.rate[1] << 8) | raw_voice.rate[0]);
-    int rate = 0;
-
-    // Gaussian interpolation using most recent 4 samples
-    int index = voice_state.fraction >> 2 & 0x3FC;
-    voice_state.fraction = (voice_state.fraction & 0x0FFF) + rate;
-    const int16_t* table  = (int16_t const*) ((char const*) gauss + index);
-    const int16_t* table2 = (int16_t const*) ((char const*) gauss + (255 * 4 - index));
-    int sample = ((table[0] * voice_state.samples[3]) >> 12) +
-                 ((table[1] * voice_state.samples[2]) >> 12) +
-                 ((table2[1] * voice_state.samples[1]) >> 12);
-    sample = (int16_t) (sample * 2);
-    sample += (table2[0] * voice_state.samples[0]) >> 11 & ~1;
-    return clamp_16(sample);
-}
 
 // Base normal_gauss table is almost exactly (with an error of 0 or -1 for each entry):
 // int normal_gauss[512];
